refactor: extract AbstractAuroraReplicationCluster from AuroraReplicationCluster
Splits the reusable scratch-database lifecycle, replica-command execution,
and status-polling logic out of AuroraReplicationCluster into a new
abstract base class, leaving only the Postgres-specific SQL and
DROP DATABASE statement in the concrete subclass. This makes it possible
to add sibling implementations for other Aurora engines without
duplicating the shared control flow.
