+++ qa/acceptance-tests/src/test/java/io/camunda/it/rdbms/db/util/AuroraReplicationCluster.java
@@ -7,264 +7,32 @@
  */
 package io.camunda.it.rdbms.db.util;
 
-import static org.awaitility.Awaitility.await;
-
-import io.camunda.zeebe.util.Unit;
-import java.io.IOException;
-import java.io.InputStream;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.time.Duration;
-import java.util.UUID;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
- * {@link ReplicationClusterContainer} backed by a pre-provisioned AWS Aurora Global Database
- * instead of local containers.
+ * {@link AbstractAuroraReplicationCluster} for an Aurora <b>PostgreSQL</b> Global Database.
  *
- * <p>An Aurora cluster cannot be spawned in CI; it is assumed to be set up beforehand and reachable
- * via the following environment variables:
- *
- * <ul>
- *   <li>{@value #ENV_JDBC_URL} – JDBC URL of the primary cluster endpoint (any maintenance
- *       database, e.g. {@code jdbc:postgresql://<endpoint>:5432/postgres})
- *   <li>{@value #ENV_USERNAME} – database user with {@code CREATEDB} privilege
- *   <li>{@value #ENV_PASSWORD} – password of that user
- *   <li>{@value #ENV_STOP_REPLICA_CMD} – shell command that removes the replica (e.g. a {@code
- *       terraform apply} scaling the secondary instance count to 0)
- *   <li>{@value #ENV_START_REPLICA_CMD} – shell command that restores the replica
- * </ul>
- *
- * <p>{@link #start()} creates a uniquely named scratch database on the cluster so each test run is
- * isolated; {@link #stop()} drops it again, leaving the cluster in a clean state. Replica removal
- * and recovery are delegated to the infrastructure-provided commands, since a managed Aurora Global
- * Database cannot be manipulated directly from a test; after {@link #startReplica()} the cluster
- * waits until the replica reports a position via {@code aurora_global_db_instance_status()}.
+ * <p>The admin JDBC URL (see {@value AbstractAuroraReplicationCluster#ENV_JDBC_URL}) is expected to
+ * be a {@code jdbc:postgresql://<endpoint>:5432/postgres}-style URL.
  */
-public final class AuroraReplicationCluster implements ReplicationClusterContainer {
-
-  public static final String ENV_JDBC_URL = "TEST_AURORA_JDBC_URL";
-  public static final String ENV_USERNAME = "TEST_AURORA_USERNAME";
-  public static final String ENV_PASSWORD = "TEST_AURORA_PASSWORD";
-  public static final String ENV_STOP_REPLICA_CMD = "TEST_AURORA_STOP_REPLICA_CMD";
-  public static final String ENV_START_REPLICA_CMD = "TEST_AURORA_START_REPLICA_CMD";
-
-  private static final Duration REPLICA_COMMAND_TIMEOUT = Duration.ofMinutes(30);
-
-  private static final Logger LOG = LoggerFactory.getLogger(AuroraReplicationCluster.class);
-
-  private final String adminJdbcUrl = requireEnv(ENV_JDBC_URL);
-  private final String username = requireEnv(ENV_USERNAME);
-  private final String password = requireEnv(ENV_PASSWORD);
-  private final String stopReplicaCommand = requireEnv(ENV_STOP_REPLICA_CMD);
-  private final String startReplicaCommand = requireEnv(ENV_START_REPLICA_CMD);
-  private final String databaseName = "camunda_it_" + UUID.randomUUID().toString().replace("-", "");
-  private final ExecutorService processExecutor =
-      Executors.newSingleThreadExecutor(r -> new Thread(r, "aurora-cluster"));
+public final class AuroraReplicationCluster extends AbstractAuroraReplicationCluster {
 
   @Override
-  public void start() {
-    LOG.info("Creating scratch database '{}' on Aurora cluster", databaseName);
-    executeAdminStatement("CREATE DATABASE " + databaseName);
+  protected String dropDatabaseStatement(final String databaseName) {
+    return "DROP DATABASE IF EXISTS " + databaseName + " WITH (FORCE)";
   }
 
   @Override
-  public void stop() {
-    LOG.info("Dropping scratch database '{}' on Aurora cluster", databaseName);
-    executeAdminStatement("DROP DATABASE IF EXISTS " + databaseName + " WITH (FORCE)");
+  protected String replicaUnreachableQuery() {
+    return """
+        SELECT server_id, session_id, durable_lsn
+        FROM aurora_global_db_instance_status()
+        """;
   }
 
   @Override
-  public String getJdbcUrl() {
-    // replace the maintenance database in the admin URL with the scratch database
-    final int lastSlash = adminJdbcUrl.lastIndexOf('/');
-    final int queryStart = adminJdbcUrl.indexOf('?', lastSlash);
-    final String query = queryStart < 0 ? "" : adminJdbcUrl.substring(queryStart);
-    return adminJdbcUrl.substring(0, lastSlash + 1) + databaseName + query;
-  }
-
-  @Override
-  public String getUsername() {
-    return username;
-  }
-
-  @Override
-  public String getPassword() {
-    return password;
-  }
-
-  @Override
-  public Future<Void> stopReplica() {
-    LOG.info("Triggering Aurora replica outage via external command");
-    final var future = executeReplicaCommand(stopReplicaCommand);
-    waitForReplicaUnreachable();
-    LOG.info("Aurora replica is unreachable");
-    return future;
-  }
-
-  @Override
-  public Future<Void> startReplica() {
-    LOG.info("Restoring Aurora replica via external command");
-    final var future = executeReplicaCommand(startReplicaCommand);
-    waitForReplication();
-    LOG.info("Aurora replica is in sync with primary");
-    return future;
-  }
-
-  /**
-   * Waits until no replica instance reports a position on the primary, i.e. the secondary has
-   * actually gone offline. The stop command (e.g. a reboot) returns before the instance is down, so
-   * without this the test would proceed while the replica is still reachable and exporting would
-   * not yet be paused.
-   */
-  private void waitForReplicaUnreachable() {
-    await()
-        .atMost(Duration.ofMinutes(5))
-        .pollInterval(Duration.ofSeconds(5))
-        .ignoreExceptions()
-        .untilAsserted(
-            () -> {
-              try (final Connection connection =
-                      DriverManager.getConnection(adminJdbcUrl, username, password);
-                  final Statement statement = connection.createStatement();
-                  final ResultSet rs =
-                      statement.executeQuery(
-                          """
-                          SELECT server_id, session_id, durable_lsn
-                          FROM aurora_global_db_instance_status()
-                          """)) {
-                int replicaCount = 0;
-                while (rs.next()) {
-                  final String sessionId = rs.getString("session_id");
-                  if (!"MASTER_SESSION_ID".equals(sessionId)) {
-                    replicaCount++;
-                    LOG.info(
-                        "Replica still reachable: server_id={}, session_id={}, durable_lsn={}",
-                        rs.getString("server_id"),
-                        sessionId,
-                        rs.getLong("durable_lsn"));
-                  }
-                }
-                if (replicaCount > 0) {
-                  throw new AssertionError(
-                      "Replica still reachable (replicas visible: " + replicaCount + ")");
-                }
-              }
-            });
-    LOG.info("Replica is unreachable now");
-  }
-
-  private Future<Void> executeReplicaCommand(final String command) {
-    LOG.info("Executing replica command: {}", command);
-    try {
-      final Process process =
-          new ProcessBuilder("sh", "-c", command).redirectErrorStream(true).start();
-      final String output;
-      try (final InputStream is = process.getInputStream()) {
-        output = new String(is.readAllBytes());
-        return processExecutor.submit(
-            () -> {
-              try {
-                final boolean finished =
-                    process.waitFor(REPLICA_COMMAND_TIMEOUT.toMillis(), TimeUnit.MILLISECONDS);
-                if (!finished) {
-                  process.destroyForcibly();
-                  LOG.error("Replica command timed out. Output:\n{}", output);
-                  throw new IllegalStateException("Replica command timed out: " + command);
-                }
-                final int exitCode = process.exitValue();
-                LOG.info(
-                    "Replica command finished with exit code {}. Output:\n{}", exitCode, output);
-                if (exitCode != 0) {
-                  throw new IllegalStateException(
-                      "Replica command failed with exit code %d: %s".formatted(exitCode, command));
-                }
-                return Unit.unit();
-              } catch (final InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new IllegalStateException(
-                    "Interrupted while running replica command: " + command, e);
-              }
-            });
-      }
-    } catch (final IOException e) {
-      throw new IllegalStateException("Failed to run replica command: " + command, e);
-    }
-  }
-
-  /** Waits until a replica instance reports a replicated position on the primary endpoint. */
-  private void waitForReplication() {
-    await()
-        .atMost(Duration.ofMinutes(45))
-        .pollInterval(Duration.ofSeconds(30))
-        .ignoreExceptions()
-        .untilAsserted(
-            () -> {
-              try (final Connection connection =
-                      DriverManager.getConnection(adminJdbcUrl, username, password);
-                  final Statement statement = connection.createStatement();
-                  final ResultSet rs =
-                      statement.executeQuery(
-                          """
-                          SELECT server_id, session_id, aws_region, durable_lsn, visibility_lag_in_msec
-                          FROM aurora_global_db_instance_status()
-                          """)) {
-                int replicaCount = 0;
-                boolean replicaInSync = false;
-                while (rs.next()) {
-                  final String sessionId = rs.getString("session_id");
-                  final long durableLsn = rs.getLong("durable_lsn");
-                  // NULL session_id is treated as non-master (a recovering secondary may report a
-                  // NULL session before it establishes a streaming session); equals() is NULL-safe.
-                  final boolean isMaster = "MASTER_SESSION_ID".equals(sessionId);
-                  LOG.info(
-                      "Instance status: server_id={}, session_id={}, aws_region={}, durable_lsn={}, visibility_lag_in_msec={}",
-                      rs.getString("server_id"),
-                      sessionId,
-                      rs.getString("aws_region"),
-                      durableLsn,
-                      rs.getLong("visibility_lag_in_msec"));
-                  if (!isMaster) {
-                    replicaCount++;
-                    if (durableLsn > 0) {
-                      replicaInSync = true;
-                    }
-                  }
-                }
-                if (replicaInSync) {
-                  return;
-                }
-                throw new AssertionError(
-                    "No replica instance reporting a durable_lsn yet (replicas visible: "
-                        + replicaCount
-                        + ")");
-              }
-            });
-  }
-
-  private void executeAdminStatement(final String sql) {
-    try (final Connection connection =
-            DriverManager.getConnection(adminJdbcUrl, username, password);
-        final Statement statement = connection.createStatement()) {
-      statement.execute(sql);
-    } catch (final SQLException e) {
-      throw new IllegalStateException("Failed to execute on Aurora cluster: " + sql, e);
-    }
-  }
-
-  private static String requireEnv(final String name) {
-    final String value = System.getenv(name);
-    if (value == null || value.isBlank()) {
-      throw new IllegalStateException("Missing required environment variable: " + name);
-    }
-    return value;
+  protected String replicaStatusQuery() {
+    return """
+        SELECT server_id, session_id, aws_region, durable_lsn, visibility_lag_in_msec
+        FROM aurora_global_db_instance_status()
+        """;
   }
 }
